+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The releases of Garlmap aim to follow [semantic versioning](https://semver.org).
 ### Changed
 
 - Failed lyrics requests are now logged in the console so they can be inspected
+- Only valid fields are now used to split strings in a query, making it possible search for : and = in fields
+- Casing in field value now also makes it case sensitive
 
 ### Fixed
 
@@ -30,7 +32,7 @@ The releases of Garlmap aim to follow [semantic versioning](https://semver.org).
 
 ### Versions
 
-- Electron 25.0.0-beta.5 (was 24.1.1)
+- Electron 25.0.0-beta.7 (was 24.1.1)
 - Chromium 114.0.5735.16 (was 112.0.5615.50)
 
 ## [1.5.0](https://github.com/Jelmerro/Garlmap/compare/1.4.0...1.5.0) - 2023-04-14